requirements.txt
mysql-connector-python
sqlalchemy
pymysql
pandas
matplotlib
geopandas
numpy
shapely
